Fixes due to ScoreSaber and BeatSaver API changes, changed optional for classes to start initialized empty, always check error string

+++ src/Types/BeatSaver/UserStats.cpp
@@ -7,7 +7,7 @@ DESERIALIZE_METHOD(BeatSaver, UserStats,
     DESERIALIZE_VALUE(RankedMaps, rankedMaps, Int)
     DESERIALIZE_VALUE(AvgBpm, avgBpm, Float)
     DESERIALIZE_VALUE(AvgScore, avgScore, Float)
-    DESERIALIZE_VALUE(AvgDuration, avgDuration, Int)
+    DESERIALIZE_VALUE(AvgDuration, avgDuration, Float)
     DESERIALIZE_VALUE(FirstUpload, firstUpload, String)
     DESERIALIZE_VALUE(LastUpload, lastUpload, String)
     DESERIALIZE_CLASS(DiffStats, diffStats)
